+++ models/ride.mjs
@@ -20,6 +20,7 @@ import {
   RIDE_MODEL_NAME,
   USER_MODEL_NAME,
 } from './helpers/constants';
+import { compareTokens, getClientURL } from './helpers/custom-methods';
 
 const DEFAULT_TIMEZONE = config.get('default_timezone');
 const { DateTime, Duration } = Luxon;
@@ -69,7 +70,7 @@ const RideSchema = new Schema({
     },
   },
   departure: {
-    _id: { type: String, required: true, alias: 'departure.id' },
+    _id: { type: String, alias: 'departure.id' },
     label: String,
     location: {
       type: {
@@ -82,7 +83,7 @@ const RideSchema = new Schema({
     },
   },
   arrival: {
-    _id: { type: String, required: true, alias: 'arrival.id' },
+    _id: { type: String, alias: 'arrival.id' },
     label: String,
     location: {
       type: {
@@ -241,7 +242,6 @@ RideSchema.statics.formatFilters = function formatFilters(rawFilters, queryFilte
   delete filter.start;
   delete filter.end;
 
-
   if (filter.current) {
     let status;
     if (filter.current === 'false') {
@@ -355,13 +355,9 @@ RideSchema.methods.sendSMS = async function sendUserSMS(body) {
   return null;
 };
 
-RideSchema.methods.compareTokens = function compareTokens(token) {
-  return this.token && token && this.token === token;
-};
+RideSchema.methods.compareTokens = compareTokens;
 
-RideSchema.methods.getRideClientURL = function getRideClientURL() {
-  return `${config.get('user_website_url')}/${this.id}?token=${this.token}`;
-};
+RideSchema.methods.getClientURL = getClientURL;
 
 RideSchema.methods.getSatisfactionQuestionnaireURL = function getSatisfactionQuestionnaireURL() {
   return `${config.get('user_website_url')}/rating?rideId=${this.id}`;