+++ models/time-slot.mjs
@@ -1,12 +1,15 @@
 import mongoose from 'mongoose';
 import Luxon from 'luxon';
 import createdAtPlugin from './helpers/created-at';
-import { TIME_SLOT_COLLECTION_NAME, TIME_SLOT_DASHED_NAME, TIME_SLOT_MODEL_NAME } from './helpers/constants';
+import {
+  TIME_SLOT_COLLECTION_NAME,
+  TIME_SLOT_DASHED_NAME,
+  TIME_SLOT_MODEL_NAME,
+  MONTHLY, WEEKLY,
+} from './helpers/constants';
+import { countDocumentsWithin, filtersWithin, findWithin } from './helpers/custom-methods';
 
 const { DateTime } = Luxon;
-export const WEEKLY = 'weekly';
-export const MONTHLY = 'monthly';
-
 const { Schema } = mongoose;
 
 const TimeSlotSchema = new Schema({
@@ -67,35 +70,9 @@ TimeSlotSchema.pre('validate', async function preValidate(next) {
 });
 
 TimeSlotSchema.statics.getDashedName = () => TIME_SLOT_DASHED_NAME;
-
-TimeSlotSchema.statics.filtersWithin = function filtersWithin(after, before, f = {}) {
-  const filters = f;
-  filters.$or = [
-    {
-      start: {
-        $lt: before,
-      },
-      end: {
-        $gt: after,
-      },
-    },
-  ];
-  return filters;
-};
-
-TimeSlotSchema.statics.findWithin = function findWithin(after, before, filters = {}, ...rest) {
-  return this.find(
-    this.filtersWithin(after, before, filters),
-    ...rest,
-  );
-};
-
-TimeSlotSchema.statics.countDocumentsWithin = function countDocumentsWithin(after, before, filters = {}, ...rest) {
-  return this.countDocuments(
-    this.filtersWithin(after, before, filters),
-    ...rest,
-  );
-};
+TimeSlotSchema.statics.filtersWithin = filtersWithin;
+TimeSlotSchema.statics.findWithin = findWithin;
+TimeSlotSchema.statics.countDocumentsWithin = countDocumentsWithin;
 
 TimeSlotSchema.methods.createNextHop = async function createNextHop() {
   if (!this.recurrence.enabled || (this.recurrence.nextHop || {})._id) {