+++ models/user.mjs
@@ -118,6 +118,11 @@ const UserSchema = new Schema({
       default: false,
     },
   },
+  licences: [{
+    type: String,
+    enum: ['B', 'C', 'D'],
+    default: 'B',
+  }],
   gprd: {
     type: Date,
     validate: {